LayoutTests/fast/xpath/nsresolver-bad-object-expected.txt

CONSOLE MESSAGE: line 0: XPathNSResolver does not have a lookupNamespaceURI method.
SUCCESS
